Update moved annotations to config
