+++ src/main/java/com/example/eventplanner/config/WebSupportConfiguration.java
@@ -1,0 +1,11 @@
+package com.example.eventplanner.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.web.config.EnableSpringDataWebSupport;
+
+import static org.springframework.data.web.config.EnableSpringDataWebSupport.PageSerializationMode.VIA_DTO;
+
+@Configuration
+@EnableSpringDataWebSupport(pageSerializationMode = VIA_DTO)
+public class WebSupportConfiguration {
+}
