+++ src/main/java/com/example/eventplanner/controllers/serviceproduct/ProductController.java
@@ -6,7 +6,6 @@ import com.example.eventplanner.model.serviceproduct.Product;
 import com.example.eventplanner.model.serviceproduct.ServiceProductCategory;
 import com.example.eventplanner.services.serviceproduct.ProductService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.data.web.config.EnableSpringDataWebSupport;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -14,12 +13,9 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-import static org.springframework.data.web.config.EnableSpringDataWebSupport.PageSerializationMode.VIA_DTO;
-
 @RestController
 @RequestMapping("/api/products")
 @RequiredArgsConstructor
-@EnableSpringDataWebSupport(pageSerializationMode = VIA_DTO)
 public class ProductController {
     private final ProductService productService;
