+++ src/main/java/com/example/eventplanner/controllers/serviceproduct/ServiceController.java
@@ -6,7 +6,6 @@ import com.example.eventplanner.model.serviceproduct.ServiceProductCategory;
 import com.example.eventplanner.services.serviceproduct.ServiceService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
-import org.springframework.data.web.config.EnableSpringDataWebSupport;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -15,12 +14,9 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 
-import static org.springframework.data.web.config.EnableSpringDataWebSupport.PageSerializationMode.VIA_DTO;
-
 @RestController
 @RequestMapping("/api/services")
 @RequiredArgsConstructor()
-@EnableSpringDataWebSupport(pageSerializationMode = VIA_DTO)
 public class ServiceController {
     private final ServiceService serviceService;
