+++ src/main/java/com/example/eventplanner/controllers/serviceproduct/ServiceProductController.java
@@ -9,7 +9,6 @@ import com.example.eventplanner.services.serviceproduct.ServiceProductService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Sort;
-import org.springframework.data.web.config.EnableSpringDataWebSupport;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -22,12 +21,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import static org.springframework.data.web.config.EnableSpringDataWebSupport.PageSerializationMode.VIA_DTO;
-
 @RestController
 @RequestMapping("/api/service-products")
 @RequiredArgsConstructor()
-@EnableSpringDataWebSupport(pageSerializationMode = VIA_DTO)
 public class ServiceProductController {
     private final ServiceProductService serviceProductService;
